Fix Pyright errors across MCP servers, add shared editor environment
Widens overly-narrow timeout params (kubectl-mcp, parallels-mcp) and
fixes real Optional-handling/typing bugs in lmstudio-code-cli (missing
None-guard before agent.mcp.tools access, untyped DEFAULTS dict, OpenAI
SDK param typing, mocked-attribute typing in tests).

Also adds a root-level editor-only venv (pyproject.toml with
package = false) covering every subproject's dependencies, since this
folder is opened as a single-folder VS Code workspace where nested
per-project .vscode/settings.json are ignored -- one shared interpreter
is what actually lets Pylance resolve imports across all of them.

+++ kubectl-mcp/kubectl_mcp/server.py
@@ -12,7 +12,7 @@ def _env() -> dict[str, str]:
     return os.environ.copy()
 
 
-async def _run(cmd: list[str], timeout: int = 60) -> str:
+async def _run(cmd: list[str], timeout: float = 60) -> str:
     proc = await asyncio.create_subprocess_exec(
         *cmd,
         stdout=asyncio.subprocess.PIPE,
